@@ -4,62 +4,43 @@
   "dependencies": {
     "@apollo/client": "3.6.9",
     "@auth0/auth0-spa-js": "^2.1.2",
-    "@material-ui/core": "4.12.4",
     "graphql": "15.6.1",
     "history": "4.10.1",
     "lodash": "4.17.21",
     "pluralize": "8.0.0",
-    "ra-data-graphql-amplication": "0.0.14",
-    "react": "16.14.0",
-    "react-admin": "3.19.12",
-    "react-dom": "16.14.0",
+    "ra-data-graphql-amplication": "1.0.2",
+    "react": "^18.3.0",
+    "react-admin": "^5.1.0",
+    "react-dom": "^18.3.0",
     "react-router-dom": "^5.3.3",
-    "react-scripts": "5.0.0",
-    "sass": "^1.39.0",
-    "web-vitals": "1.1.2"
-  },
-  "overrides": {
-    "react-scripts": {
-      "@svgr/webpack": "6.5.1"
-    }
+    "sass": "^1.39.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
+    "type-check": "tsc --noEmit",
+    "lint": "eslint --fix --ext .js,.jsx,.ts,.tsx ./src",
+    "format": "prettier --write ./src",
     "package:container": "docker build ."
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
   "devDependencies": {
-    "@testing-library/jest-dom": "5.14.1",
-    "@testing-library/react": "11.2.7",
-    "@testing-library/user-event": "13.2.0",
     "@types/history": "^4.7.11",
-    "@types/jest": "26.0.16",
     "@types/lodash": "4.14.178",
-    "@types/node": "12.20.16",
-    "@types/react": "16.14.11",
-    "@types/react-dom": "17.0.0",
+    "@types/node": "^20.10.7",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "@types/react-router-dom": "5.3.2",
+    "@typescript-eslint/eslint-plugin": "^5.60.1",
+    "@typescript-eslint/parser": "^5.60.1",
+    "@vitejs/plugin-react": "^4.0.1",
+    "eslint": "^8.43.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "prettier": "^2.8.8",
     "type-fest": "0.13.1",
-    "typescript": "4.3.5"
+    "typescript": "^5.1.6",
+    "vite": "^4.3.9"
   }
 }
