@@ -1,6 +1,6 @@
 import * as React from "react";
-import Card from "@material-ui/core/Card";
-import CardContent from "@material-ui/core/CardContent";
+import Card from "@mui/material/Card";
+import CardContent from "@mui/material/CardContent";
 import { Title } from "react-admin";
 const Dashboard = () => (
   <Card>
