@@ -2,9 +2,9 @@ import * as React from "react";
 import { List, Datagrid, ListProps, DateField, TextField } from "react-admin";
 import Pagination from "../Components/Pagination";
 
-export const MorList = (props: ListProps): React.ReactElement => {
+export const VikaList = (props: ListProps): React.ReactElement => {
   return (
-    <List {...props} title={"Mors"} perPage={50} pagination={<Pagination />}>
+    <List {...props} title={"Vikas"} perPage={50} pagination={<Pagination />}>
       <Datagrid rowClick="show" bulkActionButtons={false}>
         <DateField source="createdAt" label="Created At" />
         <TextField label="ID" source="id" />
